feature: Refactoring and stream response
