+++ transactions-details/src/main/java/ba/klika/focus/reactive/transactions/transactionsdetails/config/KafkaConfig.java
@@ -1,20 +1,20 @@
 package ba.klika.focus.reactive.transactions.transactionsdetails.config;
 
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
 import ba.klika.focus.reactive.transactions.transactionsdetails.model.Transaction;
 import ba.klika.focus.reactive.transactions.transactionsdetails.util.TransactionDeserializer;
+import reactor.kafka.receiver.KafkaReceiver;
+import reactor.kafka.receiver.ReceiverOptions;
+import reactor.kafka.receiver.internals.ConsumerFactory;
+import reactor.kafka.receiver.internals.DefaultKafkaReceiver;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import reactor.kafka.receiver.KafkaReceiver;
-import reactor.kafka.receiver.ReceiverOptions;
-import reactor.kafka.receiver.internals.ConsumerFactory;
-import reactor.kafka.receiver.internals.DefaultKafkaReceiver;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 
 @Configuration
 public class KafkaConfig {
