+++ transactions-details/src/main/java/ba/klika/focus/reactive/transactions/transactionsdetails/resource/AccountTransactionsResource.java
@@ -1,26 +1,27 @@
 package ba.klika.focus.reactive.transactions.transactionsdetails.resource;
 
 import ba.klika.focus.reactive.transactions.transactionsdetails.model.AccountTransaction;
-import ba.klika.focus.reactive.transactions.transactionsdetails.model.AccountTransactionMapper;
 import ba.klika.focus.reactive.transactions.transactionsdetails.model.Transaction;
 import ba.klika.focus.reactive.transactions.transactionsdetails.service.AccountDetailsService;
+import ba.klika.focus.reactive.transactions.transactionsdetails.util.AccountTransactionMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
 import reactor.kafka.receiver.KafkaReceiver;
 import reactor.kafka.receiver.ReceiverRecord;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
 @RequiredArgsConstructor
-public class TransactionsResource {
+public class AccountTransactionsResource {
     private final KafkaReceiver<String, Transaction> kafkaReceiver;
     private final AccountTransactionMapper accountTransactionMapper;
     private final AccountDetailsService accountDetailsService;
 
-    @GetMapping("/accounts/transactions")
+    @GetMapping(value = "/accounts/transactions", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
     public Flux<AccountTransaction> getTransactions() {
         return kafkaReceiver.receive()
                 .checkpoint("Message consuming started")
