+++ transactions-details/src/main/java/ba/klika/focus/reactive/transactions/transactionsdetails/service/AccountDetailsService.java
@@ -1,6 +1,5 @@
 package ba.klika.focus.reactive.transactions.transactionsdetails.service;
 
-import ba.klika.focus.reactive.transactions.transactionsdetails.connector.accountdetails.AccountDetailsConnector;
 import ba.klika.focus.reactive.transactions.transactionsdetails.model.AccountTransaction;
 import lombok.RequiredArgsConstructor;
 import reactor.core.publisher.Mono;
@@ -9,12 +8,12 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class AccountDetailsService {
-    private final AccountDetailsConnector accountDetailsConnector;
+    private final AccountDetailsConnectorService accountDetailsConnectorService;
 
     public Mono<AccountTransaction> getAccountsInTransaction(AccountTransaction accountTransaction) {
         return Mono.zip(
-                accountDetailsConnector.getAccount(accountTransaction.getSourceIban()),
-                accountDetailsConnector.getAccount(accountTransaction.getDestinationIban()),
+                accountDetailsConnectorService.getAccount(accountTransaction.getSourceIban()),
+                accountDetailsConnectorService.getAccount(accountTransaction.getDestinationIban()),
                 (source, destination) -> accountTransaction
                         .setSourceAccount(source).setDestinationAccount(destination)
         );
