+++ transactions-details/src/main/java/ba/klika/focus/reactive/transactions/transactionsdetails/util/TransactionDeserializer.java
@@ -1,13 +1,13 @@
 package ba.klika.focus.reactive.transactions.transactionsdetails.util;
 
+import java.io.IOException;
+
 import ba.klika.focus.reactive.transactions.transactionsdetails.model.Transaction;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.apache.kafka.common.errors.SerializationException;
 import org.apache.kafka.common.serialization.Deserializer;
-
-import java.io.IOException;
 
 public class TransactionDeserializer implements Deserializer<Transaction> {
     private final ObjectMapper objectMapper;
